streamlit/requirements.txt
pandas
SQLAlchemy
python-dotenv
cloud-sql-python-connector[pg8000]
streamlit
pinecone-client
openai
google-cloud-storage
